added response corrections

+++ controller/admin.js
@@ -4,7 +4,28 @@ const Survey = database.surveys;
 const Role = database.roles;
 const Op = database.Sequelize.Op;
 
+const jwt = require("jsonwebtoken");
+
 exports.getlist = (req, res, next) => {
+  const adminId = req.params.id;
+
+  // if (req.headers.userid == null) {
+  //   return res.status(401).json({ msg: "Unauthorized" });
+  // }
+  // if (req.headers.token == null) {
+  //   return res.status(401).json({ msg: "Unauthorized" });
+  // }
+  // if (
+  //   req.headers.userid !=
+  //   JSON.parse(atob(req.headers.token.split(".")[1])).userId
+  // ) {
+  //   return res.status(401).json({ msg: "Unauthorized" });
+  // }
+  // if (req.headers.userid != 1) {
+  //   return res.status(400).json({ error: "user is not permitted" });
+  // }
+
+
   User.findAll()
     .then((user) => {
       return res.json(user);
@@ -17,10 +38,27 @@ exports.getlist = (req, res, next) => {
 
 exports.getAdmin = (req, res, next) => {
   const adminId = req.params.id;
+  // if (req.headers.userid == null) {
+  //   return res.status(401).json({ msg: "Unauthorized" });
+  // }
+  // if (req.headers.token == null) {
+  //   return res.status(401).json({ msg: "Unauthorized" });
+  // }
+  // if (
+  //   req.headers.userid !=
+  //   JSON.parse(atob(req.headers.token.split(".")[1])).userId
+  // ) {
+  //   return res.status(401).json({ msg: "Unauthorized" });
+  // }
+
   User.findOne({
     where: { id: adminId },
   })
     .then((user) => {
+      if (user == null) {
+        return res.status(400).json({ msg: "admin not found" });
+      }
+
       return res.status(200).json(user);
     })
     .catch((e) => {
@@ -30,6 +68,22 @@ exports.getAdmin = (req, res, next) => {
 
 exports.delete = (req, res, next) => {
   const adminId = req.params.id;
+  // if (req.headers.userid == null) {
+  //   return res.status(401).json({ msg: "Unauthorized" });
+  // }
+  // if (req.headers.token == null) {
+  //   return res.status(401).json({ msg: "Unauthorized" });
+  // }
+  // if (req.headers.userid != 1) {
+  //   return res.status(400).json({ error: "user is not permitted" });
+  // }
+  // if (
+  //   req.headers.userid !=
+  //   JSON.parse(atob(req.headers.token.split(".")[1])).userId
+  // ) {
+  //   return res.status(401).json({ msg: "Unauthorized" });
+  // }
+
 
   Survey.update(
     {
@@ -39,10 +93,14 @@ exports.delete = (req, res, next) => {
       where: { userId: adminId },
     }
   ).then((data) => {
+    if (data == 0) {
+      return res.status(400).json({ msg: "admin not found" });
+    }
     User.destroy({
       where: { id: adminId },
     }).then((data) => {
-      return res.json({ msg: "admin removed successfully" });
+      return res.status(200).json({ msg: "admin removed successfully" });
+
     });
   });
 };
@@ -51,6 +109,20 @@ exports.udpate = (req, res, next) => {
   const adminId = req.body.id;
   const adminName = req.body.name;
   const adminEmail = req.body.email;
+
+  // if (req.headers.userid == null) {
+  //   return res.status(401).json({ msg: "Unauthorized" });
+  // }
+  // if (req.headers.token == null) {
+  //   return res.status(401).json({ msg: "Unauthorized" });
+  // }
+  // if (
+  //   req.headers.userid !=
+  //   JSON.parse(atob(req.headers.token.split(".")[1])).userId
+  // ) {
+  //   return res.status(401).json({ msg: "Unauthorized" });
+  // }
+
 
   User.update(
     {
@@ -62,6 +134,10 @@ exports.udpate = (req, res, next) => {
     }
   )
     .then((data) => {
+      if (data == 0) {
+        return res.status(400).json({ msg: "admin not found" });
+      }
+
       return res.status(200).json({ msg: "Admin Updated Successfully" });
     })
     .catch((error) => {
@@ -72,6 +148,19 @@ exports.udpate = (req, res, next) => {
 };
 
 exports.passwordupdate = (req, res, next) => {
+  // if (req.headers.userid == null) {
+  //   return res.status(401).json({ msg: "Unauthorized" });
+  // }
+  // if (req.headers.token == null) {
+  //   return res.status(401).json({ msg: "Unauthorized" });
+  // }
+  // if (
+  //   req.headers.userid !=
+  //   JSON.parse(atob(req.headers.token.split(".")[1])).userId
+  // ) {
+  //   return res.status(401).json({ msg: "Unauthorized" });
+  // }
+
   console.log(req.body);
   const adminId = req.body.id;
   User.findOne({
@@ -98,6 +187,34 @@ exports.passwordupdate = (req, res, next) => {
 };
 
 exports.register = (req, res, next) => {
+  // if (req.headers.userid == null) {
+  //   return res.status(401).json({ msg: "Unauthorized" });
+  // }
+  // if (req.headers.token == null) {
+  //   return res.status(401).json({ msg: "Unauthorized" });
+  // }
+  // if (req.headers.userid != 1) {
+  //   return res.status(400).json({ error: "user is not permitted" });
+  // }
+  // if (
+  //   req.headers.userid !=
+  //   JSON.parse(atob(req.headers.token.split(".")[1])).userId
+  // ) {
+  //   return res.status(401).json({ msg: "Unauthorized" });
+  // }
+
+  User.findOne({
+    where: { email: req.body.email },
+  })
+    .then((user) => {
+      if (user != null) {
+        return res.status(400).json({ msg: "user already exist" });
+      }
+    })
+    .catch((e) => {
+      return res.json({ msg: "Admins Fetch Failed" });
+    });
+
   Role.findOne({
     where: {
       rolename: "admin",
@@ -118,10 +235,11 @@ exports.register = (req, res, next) => {
     };
     User.create(userdata)
       .then((data) => {
-        return res.json({ msg: "user registered successfully" });
+        return res.status(200).json({ msg: "user registered successfully" });
       })
       .catch((e) => {
-        return res.json({ msg: "user registered failed" });
+        return res.status(400).json({ msg: "user registered failed" });
+
       });
   });
 };
